fix(bridge): wrap the default messages into a mutable map
This resolves an issue when converting the old configuration file. The loader tries to put all the defined messages into the default messages. This step is required to both copy the old messages which are already set and add the new messages which were added during the development of the release.