fix(bridge): Prevents disconnects when only one lobby is available
Velocity will disconnect the player if the result of KickedFromServerEvent is redirecting to the current server of the player. This commit fixed that behaviour while also sending the reason for the server disconnect to the player.
This commit also re-introduces the method PlatformBridgeManagement#getCachedService(UUID) which was removed while reverting commit 9cdd1851968cb8fb3557b65cff59e66367af7360