chore: Drop loading of remote service versions
The new updater will download the latest version of CloudNet every time the launcher starts. This prevents users from needing to change the versions.json url every time they want to use the fallback versions. With the new updater we can more easily deploy updates to the end user, removing the need of the remote services.json file. Via the rest-api the loading of remote versions is still possible.