Improve old launcher (#569)
* rework(launcher): Improve old launcher

* Update publishing step

Discard all previous commits to the target branch to keep the repository as small as possible and do not push to the target repo if no files will be affected by the push.

* chore: re-add artifact upload after build

* chore: small adjustments to build action

* chore: Drop loading of remote service versions

The new updater will download the latest version of CloudNet every time the launcher starts. This prevents users from needing to change the versions.json url every time they want to use the fallback versions. With the new updater we can more easily deploy updates to the end user, removing the need of the remote services.json file. Via the rest-api the loading of remote versions is still possible.

* chore(launcher): Drop NonNull annotation in java6 module

As the NonNull annotation makes use of ElementType.TYPE_USE which isn't present in java 6 yet, leading to compiler warnings. On the other hand we don't need the annotation at the 'main(String[] args)' as the arguments will always be non-null.

* (Hopefully) Copy plugin jars into final zip file

* Remove node from final zip as the updater should handle that