chore: disable "greload" on bungeecord and friends
Waterfall decided to all proxy reloaders to remove servers from the already registered ones. This will break CloudNet as all services are unregistered and players will be disconnected from the service (services are not written to the bungee configuration file).
Listening to the ProxyReloadEvent is not an option either as the removal of services will happen before the event is fired...

See: https://github.com/WaterfallMC/Waterfall-Old/issues/17