chore: add support for compressed http responses (#699)
### Motivation
Clients can send an `accept-encoding` http header to request a specific encoding of the response, some of them require the server to compress the response (especially useful when trying to f. ex. download files).

### Modification
Add the content compressor handler of netty into the pipeline which automatically compresses the response when requested. (Supported compressions are `gzip`, `deflate`. `zstd` & `br`). 

### Result
We are now compressing http responses when requested.