fix: disable jline logging to prevent stack overflow exceptions (#700)
### Motivation
When enabling debug logging, we're setting the root logger into the FINEST level to print all log messages into the console. This causes the jline logger to try the same, leading to stack overflow exceptions when logging into the console. (Logger.log -> console.write -> Logger.log etc.)

### Modification
Disable the logger instances of jline which are indirectly used by us (there are more, but we don't use the classes they are in).

### Result
No more exceptions when enabling debug mode.