chore: update to netty 5.0.0-Alpha3 (#709)
### Motivation
Netty has released [the next alpha version of netty 5](https://netty.io/news/2022/07/08/5-0-0-Alpha3.html). While we were using snapshot versions previously, we now want to stay in the release channel of netty to prevent accidental pulls of breaking changes.

### Modification
Bump netty to 5.0.0-Alpha3 and fix an issue with a breaking change.

### Result
We are now using the latest release version of netty 5.