chore: improve host and port parsing during setups (#713)
### Motivation
Currently host and port inputs must be in an uri format to be parsed correctly from our host and port parser. This makes the setup process much harder for users.

### Modification
Remove the need for the uri-syntax and assume the formatting of the input based on the `withPort` argument passed into the parser. The parser will still accept uri formatted input like: `[2001:db8::1]:80`. We are now also normalizing non-ascii input like `☃.net` to an ascii format (in this example case `xn--n3h.net`).

### Result
Much better parsing of host and ports during setups, making the process much easier.