fix: load wrapper network client certificates from the correct path (#716)
### Motivation
The network server certificates (which are used by the wrapper) are copied into the wrapper directory before the service starts, if the ssl configuration is enabled. This is required as the service might not have permissions to access the certificates when started (for example running in docker, but the certificates are stored outside a readable folder). On the other hand it makes no sense to copy the ssl certificates into the service directory and never use them.

### Modification
When copying the wrapper ssl certificates rewrite the ssl configuration passed to the wrapper in a way that it loads the certificates from the `.wrapper` directory.

### Result
The wrapper now loads the certificates which were copied into the service directory.