chore: allow updates of service properties (#717)
### Motivation
Some plugins might set custom properties into a service info snapshot and then update the snapshot safely to all services in the network. There is currently no way to do that, aside from the service to which the snapshot belongs.

### Modification
Added a `updateProperties` method to the SpecificCloudServiceProvider which allows safely updating the service properties by requesting a snapshot with the changes applied of the plugins running on the target service. The plugins are also responsible to decide if they want to override the properties or keep them.

### Result
An easy and safe way to update service properties.