feat: allow a host address on a per task basis (#727)
### Motivation
Currently every service is started on the same host address provided in the node configuration. The problem with this is that there is no way to use different hosts for different tasks like a Proxy and a Lobby.

### Modification
Created a new host address option for tasks which is used to bind the services of the task to. In order to make this feature available in the cluster too (different ips on different host systems), an ip alias system was implemented. The ip aliases are configured on each node and are resolved when starting the service. If somebody wants to use the same ip for each task supplying null as host address will use the fallback host, which is the old approach of using the host address provided by the node.

### Result
All tasks can have different host addresses and bind their services to that host address.