feat: allow service creation retries (#761)
### Motivation
Currently there is no way for a plugin to automatically schedule the retry of a service creation if the initial request failed. 

### Modification
Added the possibility to specify a retry configuration in the service configuration which allows the retry of a service creation. The creation has a fixed amount of max retries (there is no way to schedule forever) and the possibility to backoff the service re-creation according to the current retry count.

Once the creation process suceeded, an event will be fired on previously added listeners for the state of the service creation. The deferred creation can be associated with the service by either using the creation id of the CreationResult or by (for example) adding a marker to the supplied service configuration when creating the service initially.

### Result
API users now have a much easier way to define how a service creation should be retried.