chore: correctly mark interrupted threads as interrupted (#768)
### Motivation
The interrupted state of a thread will be reset in case an InterruptedException is thrown. As we need catch these exceptions any upper layer of the method call is unable to respond to the interuption and `Thread.currentThread().isInterrupted()` will falsely return false.

### Modification
Call `Thread.currentThread().interrupt()` to mark threads as interrupted when catching an InterruptedException. This PR also removes the usages of `Thread.interrupted()` which check if a thread is interrupted but also resets the interrupted state of the thread.

### Result
No more accidental wrong marks of threads which should be marked as interrupted.

##### Other context
A bit more context is given (for example) here: https://dzone.com/articles/why-do-we-need-threadcurrentthreadinterrupt-in-int