ci: let build and tests run on zulu-19-ea as well. (#772)
### Motivation
All tests should run on the latest lts version of java (currently 17) and the latest available version (currently 19-ea). The change of the distribution to zulu was needed as temurin had no 20-ea builds, but we're currently limited by mockito which only supports up to java 19. To make later updates easier the distribution should be zulu anyway.

### Modification
Use java 19-ea as the latest test version instead of 18, replaced temurin with zulu jdk distribution.

### Result
Faster updates to newer java versions & tests are running on jdk 19 which makes it easier to determine if we need to do something in order to support a newer java version.