chore: use sponge download api instead of hardcoded urls (#771)
### Motivation
Sponge provides a downloads api which allows to query the latest version of a sponge artifact. This is much nicer and removes the need to manually ensure that the download url of the sponge versions are always up-to-date.

### Modification
Remove hardcoded sponge download urls from versions.json and replaced them with an install step which fetches the latest version from the sponge downloads api.

### Result
Easy ensure that the downloaded version is always the latest one instead of relying on manual updates.