fix: fixup invalid command sender message (#799)
Motivation
The invalid command sender message is currently not showing the required sender as the argument was escaped by the surrounding '.

Modification
Fixup the escape of the message arguments, as well as fixing the message to correctly reflect what it is meant to show. This requires to override the toString method in all registered command senders to get a human readable name of the sender during formatting.

Result
The invalid command sender message shows up correctly.