fix: properly rewrite service configuration files (#800)
Motivation
Some configuration files use indentations at some points, for example for the host address of a service. We previously hardcoded the indention of the modified lines, but the indention of the lines don't need to be the exact same every time.
Modification
Re-use the indentation of the line which was previously applied to prevent errors from wrong indentations.
Result
The indentation of config files is preserved and no more errors occur because our hardcoded indention is off.