chore: properly unregister commands from backing manager (#803)
Motivation
Unregistering a command is currently not done in the backing cloud command manager, the command is only removed from our known registered commands (which isn't working either as the wrong class loader in association with the commands was stored). This causes exceptions, when for example a module is updated during runtime, the new file starts and tries to register the same command as previously registered by the old module.
Modification
Due to a relatively new feature in cloud it is possible to unregister commands. This PR makes now use of that feature.
Result
Unregistering a command is now working as expected and later registrations of the same command are possible.