feat: improve mysql database searches by introducing json columns (#805)
Motivation

During the testing phase of RC1 some users reported issues with their key length in SQL databases. The limit was set to 64 chars which is way to short for our purposes.
Modification

Upped the key limit to 512 chars to be safe for the future. While doing this change I noticed that we could use the JSON type and its functions to optimize our searching
Result

The limit is set to 512 chars and all (SQL) tables are using json now.
Other context