fix: resolve convert issues of old npc/sign entries (#810)
Motivation
The old npc & sign entries are not converted correctly from v3 to v4 as h2 stores all database names uppercase (and we convert them to lowercase). As the entries are stored in a database called cloudNet_module_configuration this leads to an issue when used with databases which use case-sensitive table naming (like xodus or mongodb).

Modification
Check box database spelling variants (cloudNet_module_configuration and cloudnet_module_configuration) in order to find the database which contains the old entries after converting the database from h2.
On the other hand the npcs were parsed to a set type which caused the emit of some npc entries, therefore I changed the type to a collection to read all entries from the old document.

Result
The created signs and npcs on v3 are now correctly converted to v4.