chore: replace some AtomicXXX classes with their non-atomic equivalent (#813)
### Motivation
Atomic wrappers are used to prevent concurrency issues caused by multiple threads accessing the same field. In CloudNet we use these atomic wrappers in some places, where calls to the fields (or methods which are mutating them) are only done either within a locked context or only from one thread. Therefore the atomic wrapping is not needed and can be replaced with the lighter direct accesses.

### Modification
Replace some AtomicXXX wrapper classes when they are not needed with their non-atomic equivalent.

### Result
No more atomic actions when not needed.