fix: prevent double installation of modules (#812)
### Motivation
Currently the `module install` command doesn't verify if a module is already installed (either loaded or just the file being in the modules directory). This can lead to errors when loading the module, for example because of duplicate command registrations.

### Modification
Ensure that `module install` is only able to install modules which weren't installed previously. The suggestions now only list modules which aren't loaded as well. For the suggestions this isn't perfect as the file might be available in the module directory, but reading all files and checking if the file is there is too heavy to do.

### Result
Modules can only be installed once and now need a proper uninstall before they can be installed again.