fix(sftp): properly close all ssh client connection (#820)
### Motivation
Currently the pooled sftp clients are not disconnected properly from the
remote server on close, which means that we rely on the fact that the
server will timeout the connection at some point. If no client ever
connected to a remote ssh server, an exception is thrown on shutdown as
well.

### Modification
Properly close all pooled sftp clients and only disconnect the root
client if it got initialized.

### Result
All connections are now closed on shutdown properly and no more
exception are thrown.