fix: correctly override methods from ServiceConfigurationBase#Builder (#825)
### Motivation
Some methods from ServiceConfigurationBase#Builder are not overriden in
ServiceTask#Builder but need to be in order to delegate these method
calls to the underlying ServiceConfigurationBase#Builder methods. This
results in incorrect modification of the underlying options, for example
when using `tasks task <name> add jvmOption`.

### Modification
Override builder methods in ServiceTask#Builder that need to delegate
their calls.

### Result
Correctly modify the underlying ServiceConfigurationBase.