fix: only use fallback with forced host on initial connect (#835)
### Motivation
Currently fallback configurations with a forced-host set will always be marked as a possible fallback for a player, as the proxy will keep the host which the player used to join. This is wrong as players are unable to connect to a different fallback after joining via a forced-host.

### Modification
Only use forced-host fallback configuration on initial connection.

### Result
Players can connect to a different fallback after joining using a forced host.