fix: correctly reset fallback profiles when keeping server connection (#858)
### Motivation
When a player connection to the current lobby server is kept due to a
server connection issue, the fallback profile of the player is not
cleared properly, which causes the fallback server to be ignored when
trying the connection again. This leads to the player being disconnected
from the proxy.

### Modification
Clear the fallback profile of the player when the connection to the
current server is kept open.

### Result
No more disconnects of players when the fallback connection is kept
open.