fix: correctly convert npcs and ensure correct scoreboards (#877)
### Motivation
The current migration attempt does not consider that the display name
the user used now is not displayed anymore at all. Therefore we should
take the display name and set it as first info line. Furthermore we did
not ensure that the player has a scoreboard just for himself, which
might lead to issues with other plugins working with bukkit scoreboards.

### Modification
Added the display name into the migration process. Ensured that each
player has it's own scoreboard and we don't replace any other
scoreboards

### Result
The migration is now correct and the scoreboards of other plugins are
not replaced.