fix: resolve some small issues in the current npc bukkit implementation (#890)
### Motivation
There were two smaller issues causing exceptions when using the npc
module. The first one was caused by an attempt to add a not-spawned npc
to a scoreboard team on join. The second one was caused by the npc lib,
therefore there is no real fix for that in this pr.

### Modification
* Added a check which verifies that a npc is actually spawned before
trying to add it to the scoreboard.
* Improved the handling of worlds within the npc module. Altough it
didn't cause issues yet, especially the current Chunk Un-/Loading code
is prone to cause issues on some servers.
* Added a check which verifies that changes to worlds a npc is located
in are reflected into the `location()` and `canSpawn` methods.

### Result
The npc module should no longer throw exceptions on join and should
handle world events much smoother.

##### Other context
Fixes #889