fix: resolve smart module regression causing invalid template inclusions (#910)
### Motivation
In RC3 (PR #880) there was an issue introduced causing invalid template
inclusions for services forced by the smart module. This can happen in
two different ways:
* the smart module always did pre-inclusion of templates (if
`directTemplatesAndInclusionsSetup` was enabled) as a check was missing
if the smart module is actually enabled for the task
* the smart module force-installed all templates to the service, even
when that operation would be invalid normally. This could for example
mean that a template which shouldn't get included to a static service
get's included anyway.

### Modification
* Ensure that the smart module is enabled before executing the direct
template and inclusion setup
* Only force-install templates when the service is started for the first
time (default behaviour)
* Add api methods which give access to the `force` boolean of the
`includeWaitingServiceTemplates` method

### Result
The template installation process of the smart module is no longer
breaking existing setups and functions as expected.