chore: log the underlying ssh exceptions when getting a client from the pool fails (#911)
### Motivation
We've experienced some issues where all commands regarding templates
would take really long to execute. The issue was that the SFTP module
was loaded and active but not configured properly and therefore the
connection would fail each time. To make sure that users are aware of
missing or wrong credentials this change was made.

### Modification
Added a check that makes sure that we re-print all SSHException messages
instead of just ignoring them to make sure that users are aware of it.

### Result
No more wondering why some template commands are that slow.