fix: leave closing the buffer to netty when a channel becomes inactive (#921)
### Motivation
When a channel is closed and a call to `VarInt32FrameDecoder#decode`
happens, we're closing the buffer provided to the method. Netty on the
other hand will try to close the buffer when the inactive event is
called, causing a race which leads to an exception on either side (one
will have an exception because the buffer is already closed).

### Modification
Leave the buffer close operation to netty and just ignore the input
message.

### Result
No more exceptions caused by buffer drop races.