chore: calculate diff between node info updates based on the local time (#923)
### Motivation
Currently the milliseconds between each node update for the node
disconnection is calculated based on the system time of the remote node.
This can lead to issues and invalid disconnects when the time of the
remote (or local) node is not syncronized properly.

### Modification
Use the local time of the node which receives the node info update to
check if the node exceeds the soft disconnect limit. This should prevent
disconnects which are made because the time of the source and target
server are not syncronized. At the moment, the receive timestamp is only
used for the disconnect check, in the future it might get included in
command outputs or generated node reports as well.

### Result
No more invalid disconnects of nodes because of time desyncs.