chore: make our build setup less dependant on jitpack queue (#934)
### Motivation
Our current build is very dependant on jitpack build leading to build
timeouts and even unmet dependencies when there is a new commit pushed
to a repo we're dependening on.

### Modification
Replace the dependencies which are based off a branch with releases that
are statically served by jitpack.

### Result
Hopefull less build timeouts and cloudnet start failures due to deleted
artifacts.