chore: replace javers (#944)
### Motivation
The javers library has some pitfalls which weren't visible when first
evaluating and implementing the diff pretty printer. The main issue is
that there is no way to access changes inside containers, for example
when a collection element has a change in it, and more importantly there
is no way to access the old and new collection from the diff
information. Therefore pretty-printing the change in order to allow a
user to understand what changed and what action to take (apply, revert,
skip) is not possible without starting to depend on implementation
details and reflection.

### Modification
Replace the javers library with gulf and rewrite the pretty printer to
even better reflect changed objects and improve the style of it a tiny
bit (code and output wise).

### Result
Better object comparison and output print for users.