chore: improve cluster sync ux (#949)
### Motivation
The current cluster sync handling has some inconveniences which might
cause trouble for some users which are using it. First, the user has to
respond to all change accept/skip messages within 30 seconds if he
wanted that a response is send to the other node (in case a change from
the current node should get applied to the remote node). On the other
hand, when the first question was shown, the command handlers weren't
disabled correctly which would lead to command suggestions.

### Modification
Replace the query waiting for a response from the data sync (and
therefore from a user response) with a separate message. This also
removes the need to ensure that we answer all syncs if there are no
changes to make on the publising node. Furthermore the command handler
registration is now moved up, so disabling the handlers for the initial
data sync handling now works as expected.

### Result
Better user experience when using the cluster sync functionality.