fix: correctly fire root http handlers (#952)
### Motivation
HTTP handlers which are registered to the root path (`/`) are not fired
correctly. This is due to the reason that the incoming request uri can
never be blank: `[...] if none is present in the original URI, it MUST
be given as "/" (the server root)` from: [RFC2616, Section
5.1.2](https://www.rfc-editor.org/rfc/rfc2616#section-5.1.2).

### Modification
Correctly check if the request we received is send against the root path
before trimming off the leading slash.

### Result
Handlers added to the root path are called correctly.