chore: make module lifecycle commands console only (#956)
### Motivation
The modules command allows to change the lifecycle of the module. These
lifecycle changes might produce errors that are not redirected to the
command executor, but printed into the console. Therefore the commands
should not get executed in-game using the /cloud prefix

### Modification
Made all lifecycle changing commands require the ConsoleCommandSource

### Result
The module lifecycle chaning commands are only executable using the
console.

##### Other context
Fixes #915