chore: remove instance methods and old getters / setters (#992)
### Motivation
We're migrating everything to use dependency injection. The old plan was
to first upgrade all modules & plugins, then remove the old instances,
getters & setters. But due to unavoidable breaks in the wrapper which
are causing CloudNet to not be compileable until all usages are removed
anyway, we pull this step to now. This should make it easier to migrate
as well, as unknown methods are now clearly visiable.

### Modification
Remove CloudNetDriver, the instance method from Wrapper & Node and all
getters / setters which were still present in the Node and Wrapper
classes.

### Result
No more getters, setters & instances where we don't need them.