chore: migrate papi expansion to use injection (#993)
### Motivation
Migrates the papi expansion to use dependency injection (somewhat).

### Modification
Instead of using Wrapper.instance we now get the service info holder
instance from the ext injection layer. This is not the ideal use for
that case, but, it prevents us from needing to wrap the expansion into a
plugin to create an injection layer just for that purpose.

### Result
The papi expansion now uses injection.