chore: add registry for injection layers (#997)
### Motivation
Currently there is no way to bind an injection layer to a known
instance, the only way to identify the layer is by using some sort of
class which implements InjectionLayerHolder and is findable in the
current tree. This behaviour is especially bad when working with
platforms, as there is no way to indicate an injection holder (for
example with the class loader).

### Modification
Add a registry for injection layers and bind the plugin main classes and
plugin class loaders as identifiers in the platform plugin manager.

### Result
Identifiable injection layers for platform plugins.