feat: improve formatting for motds and placeholders in the player info (#964)
### Motivation
SyncProxy was not using the component converter in all platform
implementations leading to varying results for the same input.

### Modification
Made use of the new component converter in all implementations to make
sure that the results are the same on all platforms. Furthermore
improved code readability.

### Result
The components are converted using the converter and the result is the
same on all platforms.