fix: make sure to load cloudperms before Vault (#1084)
### Motivation
Vault expects any permission system that they do not support themselves
to be loaded before Vault. This is the only way to access permission
groups etc. via Vault.

### Modification
Removed the our depend on Vault and the VaultAPI and instead set a
loadbefore on Vault.

### Result
CloudPerms is loaded before Vault is and therefore permission system
related calls are working.