fix: correct some netty channel options (#1080)
### Motivation
Some of the currently set netty channel options are not correctly
applied to the boss channel which will (on the server side) listen for
incoming connections and accept them. Due to this, there were some
people reporting issues with re-starting a node instance as the node
port would be blocked by the idle state and not be re-used.

### Modification
The channel options that need to get applied to the boss channel are now
applied correctly, as well enabling tcp keepalive packets for
connections and enabling tcp fast connect.

### Result
There should no longer be issues with blocked ports, and all channel
options should now be applied correctly.