chore: improve and fixup connection handling between nodes (#1089)
### Motivation
In the current scenario there are two issues with the current node
connect handling during the initial startup:
* The node information of running nodes is not received properly during
the startup due to the listener for node updates being registered after
the connection process completed
* For each node connection that we're waiting for a new thread is
spawned that just busy waits until the connection is established

### Modification
To fix the first issue, the listener registration was just moved to the
NodeServerProvider which is constructed for the node connection process
and member injection is executed before the instance is used for the
first time. In the old scenario, the handler gets registered by the
NodeClusterNodeProvider which is constructed and first used **after**
the connection process and therefore the listener was not yet present.

The second issue is fixed by registering all node connections that we
need to wait for to a collection and just block the current thread until
either:
 * all node connections are established successfully
 * a timeout of 7 seconds is reached
This prevents excessive use of threads (depending on the number of
nodes) that the method previously used.

### Result
Connection to nodes in the cluster works again correctly without waiting
the full 7 seconds during startup, and there is no longer a thread
spawned for each node connection that we're waiting on.