feat: add event to modify a service configuration before creation (#1129)
### Motivation
There is currently no event that allows to modify a service
configuration on the node (via a module) to change the configuration
when required. This can for example be setting process parameters or
environment variables.

### Modification
Add an event (`CloudServiceConfigurationPrePrepareEvent`) that holds the
service configuration and the builder that is used to modify the service
configuration in the cloud service factory. The event is called before
the final modifications to the configuration are made (ensuring that the
service id is only taken once, adding the templates, deployments, ... of
the group components etc.). This ensures that no changes are made to the
configuration which could potentially break the whole internal service
handling.

### Result
The configuration of a service can be modified from a (head) node
module.

##### Other context
Resolves #1074