chore: add a channel initializer that only sets supported options (#1130)
### Motivation
Netty prints out a warning when trying to set channel options that are
not supported by a transport (only for options added by the (server-)
bootstrap). This is no big deal as the warning can be ignored, but there
are some users which are reporting it as an issue.

### Modification
Add a new channel initializer that holds the options that should be
added to a channel and only sets them if they are supported by the
channel that gets initialized. This is the best way to solve the issue,
as there is no other way to check if an option is supported. Although
some options are bound to epoll or kqueue, it might get over complicated
when checking based on that, especially considering the addition of the
io_uring transport that is coming into netty main with the next release.

### Result
No more warnings that there was a try to set channel options that are
not supported by the channel.